main.py ~ Blackboard update
The blackboard assignments can now be webscraped! This is by far the most userful (because most students use black board) and eaisiest ( because all assignments for every course is in one place ) update. * Keep in mind that the code may need to be tweaked becuase i've only tried it out on my assignments. *