Add web scraping script for Blackboard grades
This is a script seperate from main.py (for now) untill I get it to completetion, which will be for it to be able to webscrape completed assignments and their corresponding grades and then be able to calculate the total grade. Eventually I want it to even calculate the grade accounting for what is in any syllabus (dropping lowest assignment grade). Note: Make sure to replace the information like username/ password or "Page we want" with your own things that apply to you, and hoplfully it'll work for you.